source version version from txt file
