+++ ms_client/version.txt
@@ -1,0 +1,1 @@
+2.0
