+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = mediaserver-api-client
-version = 2.0
+version = file: ms_client/version.txt
 description = A python3 reference implementation of an UbiCast MediaServer API client
 long_description = file: README.md
 long_description_content_type = text/markdown
@@ -37,6 +37,7 @@ setup_requires=
 [options.package_data]
 ms_client =
     conf.json
+    version.txt
 
 [options.extras_require]
 dev =
